+++ packages/ng/demo/app/user/tile/user-tile.component.ts
@@ -13,6 +13,14 @@ export class DemoUserTileComponent implements OnInit {
 			code: require('!!prismjs-loader?lang=typescript!./basic/basic.component'),
 			markup: require('!!prismjs-loader?lang=markup!./basic/basic.component.html')
 		},
+		displayformat: {
+			code: require('!!prismjs-loader?lang=typescript!./displayFormat/display-format.component'),
+			markup: require('!!prismjs-loader?lang=markup!./displayFormat/display-format.component.html')
+		},
+		sizes: {
+			code: require('!!prismjs-loader?lang=typescript!./sizes/sizes.component'),
+			markup: require('!!prismjs-loader?lang=markup!./sizes/sizes.component.html')
+		},
 	};
 	constructor() { }
 	ngOnInit() {